+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.25 - 2023-12-24
+- Added color="accent" to <Tag>
+- Added custom colors to <Tag>
+
 ## 0.0.24 - 2023-12-23
 - Moved .hds-box class to index.css